# mercapleno-backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mercapleno API",
    "description": "API documentation generated automatically from routes",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
...
    "/api/admin/users/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "example": "any"
                  },
                  "apellido": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "direccion": {
                    "example": "any"
                  },
                  "fecha_nacimiento": {
                    "example": "any"
                  },
                  "id_rol": {
                    "example": "any"
                  },
                  "id_tipo_identificacion": {
                    "example": "any"
                  },
                  "numero_identificacion": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/productos/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "example": "any"
                  },
                  "precio": {
                    "example": "any"
                  },
                  "id_categoria": {
                    "example": "any"
                  },
                  "id_proveedor": {
                    "example": "any"
                  },
                  "descripcion": {
                    "example": "any"
                  },
                  "estado": {
                    "example": "any"
                  },
                  "imagen": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/movimientos/productos": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/productos": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}